Refactor for LogHistoryActivity and its tests
