+++ owncloudApp/src/androidTest/java/com/owncloud/android/settings/logs/LogHistoryActivityTest.kt
@@ -2,7 +2,9 @@
  * ownCloud Android client application
  *
  * @author Jesús Recio @jesmrec
- * Copyright (C) 2020 ownCloud GmbH.
+ * @author Juan Carlos Garrote Gascón
+ *
+ * Copyright (C) 2021 ownCloud GmbH.
  * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -26,19 +28,19 @@ import androidx.test.espresso.matcher.ViewMatchers.isDescendantOfA
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.rule.ActivityTestRule
 import com.owncloud.android.R
-import com.owncloud.android.ui.activity.LogHistoryActivity
+import com.owncloud.android.presentation.ui.settings.LogHistoryActivity
 import org.hamcrest.Matchers.allOf
 import org.junit.Rule
 import org.junit.Test
 
-class OCSettingsLogTest {
+class LogHistoryActivityTest {
 
     @Rule
     @JvmField
-    val activityRule = ActivityTestRule(LogHistoryActivity::class.java, true, true)
+    val activityRule = ActivityScenarioRule(LogHistoryActivity::class.java)
 
     @Test
     fun itemsToolbar() {
