+++ owncloudApp/src/main/AndroidManifest.xml
@@ -209,7 +209,7 @@
             android:name=".ui.activity.ErrorsWhileCopyingHandlerActivity"
             android:theme="@style/Theme.ownCloud" />
         <activity
-            android:name=".ui.activity.LogHistoryActivity"
+            android:name=".presentation.ui.settings.LogHistoryActivity"
             android:configChanges="orientation|screenSize" />
         <activity android:name=".ui.errorhandling.ErrorShowActivity" />
         <activity
